Move blog entities to common module and add Flyway
- Move all JPA entity/embeddable/enum files from blog module to
  common/src/main/java/com/study/common/entity/ package
- Update package declarations from com.study.blog.post/comment/entity
  to com.study.common.entity in all moved files
- Remove now-redundant same-package imports within moved files
- Update all blog module files (repositories, services, controllers,
  DTOs, tests) to import entities from com.study.common.entity
- Add Flyway dependencies to blog/build.gradle:
  flyway-core and flyway-database-postgresql:10.10.0
- Create blog/src/main/resources/db/migration/ and copy V1__init_schema.sql
- Fix misleading SQL comment: post_id on comments table DOES have a FK
  constraint (REFERENCES posts(id) ON DELETE CASCADE)
