+++ blog/build.gradle
@@ -7,5 +7,7 @@ dependencies {
     implementation project(':contract')
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
+    implementation 'org.flywaydb:flyway-core'
+    implementation 'org.flywaydb:flyway-database-postgresql:10.10.0'
     testRuntimeOnly 'com.h2database:h2'
 }
