+++ blog/src/main/java/com/study/blog/admin/AdminService.java
@@ -5,12 +5,12 @@ import com.study.blog.admin.dto.AdminStatsResponse;
 import com.study.blog.comment.CommentRepository;
 import com.study.blog.common.exception.BlogErrorCode;
 import com.study.blog.common.exception.BlogException;
-import com.study.blog.post.Post;
 import com.study.blog.post.PostLikeRepository;
 import com.study.blog.post.PostRepository;
-import com.study.blog.post.PostStatus;
-import com.study.blog.post.PostTag;
 import com.study.blog.post.PostTagRepository;
+import com.study.common.entity.Post;
+import com.study.common.entity.PostStatus;
+import com.study.common.entity.PostTag;
 import java.util.List;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
