+++ blog/src/main/java/com/study/blog/comment/CommentRepository.java
@@ -1,5 +1,6 @@
 package com.study.blog.comment;
 
+import com.study.common.entity.Comment;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
