+++ blog/src/main/java/com/study/blog/post/PostBookmarkRepository.java
@@ -1,5 +1,7 @@
 package com.study.blog.post;
 
+import com.study.common.entity.PostBookmark;
+import com.study.common.entity.PostBookmarkId;
 import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
