+++ blog/src/main/java/com/study/blog/post/PostRepository.java
@@ -1,5 +1,6 @@
 package com.study.blog.post;
 
+import com.study.common.entity.Post;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
