+++ blog/src/main/java/com/study/blog/post/PostService.java
@@ -6,6 +6,11 @@ import com.study.blog.post.dto.PostCreateRequest;
 import com.study.blog.post.dto.PostResponse;
 import com.study.blog.post.dto.PostSummaryResponse;
 import com.study.blog.post.dto.PostUpdateRequest;
+import com.study.common.entity.Post;
+import com.study.common.entity.PostBookmark;
+import com.study.common.entity.PostLike;
+import com.study.common.entity.PostStatus;
+import com.study.common.entity.PostTag;
 import java.util.List;
 import java.util.UUID;
 import org.springframework.data.domain.Page;
