+++ blog/src/main/resources/db/migration/V1__init_schema.sql
