+++ blog/src/test/java/com/study/blog/integration/AdminApiTest.java
@@ -8,14 +8,14 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.study.blog.BlogTestApplication;
-import com.study.blog.comment.Comment;
+import com.study.common.entity.Comment;
 import com.study.blog.comment.CommentRepository;
-import com.study.blog.post.Post;
-import com.study.blog.post.PostLike;
+import com.study.common.entity.Post;
+import com.study.common.entity.PostLike;
 import com.study.blog.post.PostLikeRepository;
 import com.study.blog.post.PostRepository;
-import com.study.blog.post.PostStatus;
-import com.study.blog.post.PostTag;
+import com.study.common.entity.PostStatus;
+import com.study.common.entity.PostTag;
 import com.study.blog.post.PostTagRepository;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
