+++ common/src/main/java/com/study/common/entity/BlogTest.java
@@ -1,4 +1,4 @@
-package com.study.blog.entity;
+package com.study.common.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
