+++ common/src/main/java/com/study/common/entity/Post.java
@@ -1,4 +1,4 @@
-package com.study.blog.post;
+package com.study.common.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
